Remove unused code for CRAN remotes
CRAN remotes short circuit install_remotes, so we don't actually need
all of the methods.

We can't use install_remotes normal behavior because you can't install
binaries on Windows once they are excreted, due to limitations in base
R.

+++ R/install-remote.R
@@ -27,6 +27,13 @@ install_remote <- function(remote, ..., force = FALSE, quiet = FALSE) {
     return(invisible(FALSE))
   }
 
+  if (inherits(remote, "cran_remote")) {
+    install_packages(
+      package_name, repos = remote$repos, type = remote$pkg_type,
+      ..., quiet = quiet)
+    return(invisible(TRUE))
+  }
+
   bundle <- remote_download(remote, quiet = quiet)
   on.exit(unlink(bundle), add = TRUE)
